+++ src/ATK/utilities/requests.py
@@ -13,7 +13,7 @@ def send_request(survey: str, url: str, method: str = "GET", **kwargs) -> Respon
 
     s = Session()
     retries = Retry(total=5, backoff_factor=1, status_forcelist=[500, 502, 503, 504])
-    s.mount("http://", HTTPAdapter(max_retries=retries))
+    # s.mount("http://", HTTPAdapter(max_retries=retries))
     s.mount("https://", HTTPAdapter(max_retries=retries))
 
     try: