+++ testing/lightcurve_testing.py
@@ -1,11 +1,6 @@
-import numpy as np
+import ATK.Tools as ATK
 
-from ATK.structures.definitions import Lightcurve
+SOURCE = 587316166180416640
 
-mjd = np.array([1, 2, 3])
-flux = np.array([1, 2, 3])
-flux_err = np.array([1, 2, 3])
-
-lc = Lightcurve("ztf", "g", mjd=mjd, flux=flux, flux_err=flux_err)
-
-print(lc.__dict__)
+data = ATK.query("lightcurve", survey="ztf", target=SOURCE)
+# data.show()